[issue 73] Issue with Terraform PLAN
Plan was showing a diff even if nothing was changed in the .tf file
Fixing this issue by introducing another property in the resource_configuration called resource_state
'resource_state' is a computed property and it will have the updated state of the
deployment. The 'configuration' that was previously also used for detailed view of resource machines
will only behave like an input.

Signed-off-by: Prativa Bawri <bawrip@vmware.com>