[73] Issue with Terraform PLAN
One of the reason for issue in terraform plan is the lease_days.
When creating deployment resource with lease_days(say, 10), the state file is updated with
lease_days 9(this is calculated from lease_end - lease start time). In the following iterations, even
if nothing changes in the config file, the plan always detects a change as state file has lease_days 9
and config has 10.

Fixing this by mimicing the same behavior as in the vRA UI.

`lease_days` will be used to specify the number of lease days only when the deployment is created the
first time. The `lease_end` and `lease_start` will be computed and updated in the state file.
Any update of the property lease_days will be ignored.
To update the lease, use `lease_end` property. Specify in the YYYY-MM-DD HH:MM format.

Signed-off-by: Prativa Bawri <bawrip@vmware.com>