[95] Scale out/in not working due to drift in Terraform state file (#100)
After the scale-out, the terraform state was not updated correctly.
This has been fixed by using another API to read the deployment state.

This API is only available in vRA 7.5 and above. After this change, the terraform state will not be
updated for users on vra 7.4. 
For vRA 7.4 users, it is recommended to stay on the provider version v2.0.1

Signed-off-by: Prativa Bawri <bawrip@vmware.com>